fscache: Add the synchronous waiting mechanism for the volume unhash in erofs ondemand mode
Offering: HULK
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IB5UKT

--------------------------------

This patch adds the synchronous waiting mechanism for the volume unhash in
erofs ondemand mode. There are 2 main reasons:

1) Currently volume's unhash relies on its own reference count to be 0,
partly depending on the user state release fd. That is, if the user state
fails to turn off fd properly, then the volume cannot be unhashed. Next
time when mount the specified volume with the same name, the system will
waits for the release of the previous volume. However, the write semaphore
of the corresponding superblock is not released. When traversing the sb,
hung task will occur because the read semaphore cannot be obtained:

 [mount]					[sync]
vfs_get_super                                 ksys_sync
  sget_fc                                       iterate_supers
    alloc_super
      down_write_nested ----pin sb->s_umount
    list_add_tail(&s->s_list, &super_blocks)
  erofs_fc_fill_super                             super_lock
    ...                                             down_read ----hungtask
    fscache_hash_volume ----wait for volume

2) During the umount process, object generates the cache directory entry
(cachefiles_commit_tmpfile), but it is not synchronized. After umount is
complete, the user may see that the cache directory is not generated. When
inuse() is called in user mode, the cache directory cannot be found.

The solution:
1) For the erofs on-demand loading scenario, "n_hash_cookies" has been
introduced in "struct fscache_volume", increased whenever there is a child
cookie hashed and decreased if the cookie has unhashed. When it returns
zero, the volume is awakened with unhash. FSCACHE_CACHE_SYNC_VOLUME_UNHASH
flag is introduced in "struct fscache_cache", which is used to indicate
whether this feature is enabled.
2) cachefiles_free_volume() need to be called to ensure the next mount
successful, otherwise -ENODATA will be returned because cache_priv will not
be created in new volume and the object may not be initialized. To prevent
use-after-free issue caused by the kfree(volume)/kfree(cache), "ref" and
"lock" have been introduced in "struct cachefiles_volume".

There are three benefits to this:
1) The unhash of volume does not depend on whether the fd is handled
correctly by the userland. If fd is not closed after umount, it does not
matter if it continues to be used, because object->file is already NULL as
cachefiles_clean_up_object() is called before fscache_unhash_cookie().
2) The cache directory can be guaranteed to be generated before the umount
completes unless the process is interruped. This is because
cachefiles_commit_tmpfile() is called before fscache_unhash_cookie().
3) Before this patch, it is possible that after umount, calling inuse() on
a cache entry may still shows -EBUSY, because the order of umount and
cachefiles_put_directory() is indeterminate. Now thanks to volume->lock,
calling cull/inuse after umount is finished will not return an error.

Fixes: 62ab63352350 ("fscache: Implement volume registration")
Signed-off-by: Zizhi Wo <wozizhi@huawei.com>