hisi_acc_vfio_pci: fix migration DFX doubled free and use-after-free
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IBNXQ3
CVE: NA

----------------------------------------------------------------------

According to the analysis of the current debugfs, debug_migf
in debugfs is released when it exits. The system determines
whether debug_migf is empty to prevent repeated release.
However, after the release ,it is not assigned NULL, so
the null operation is added.

Fixes: 3344c0cc132d ("hisi-acc-vfio-pci:add DFX for acc migration driver")
Signed-off-by: Longfang Liu <liulongfang@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>