KVM: x86: Snapshot the host's DEBUGCTL in common x86
mainline inclusion
from mainline-v6.14-rc6
commit fb71c795935652fa20eaf9517ca9547f5af99a76
bugzilla: https://gitee.com/openeuler/kernel/issues/ICSPRX

----------------------------------------

commit fb71c795935652fa20eaf9517ca9547f5af99a76 upstream

Move KVM's snapshot of DEBUGCTL to kvm_vcpu_arch and take the snapshot in
common x86, so that SVM can also use the snapshot.

Opportunistically change the field to a u64.  While bits 63:32 are reserved
on AMD, not mentioned at all in Intel's SDM, and managed as an "unsigned
long" by the kernel, DEBUGCTL is an MSR and therefore a 64-bit value.

Reviewed-by: Xiaoyao Li <xiaoyao.li@intel.com>
Cc: stable@vger.kernel.org
Reviewed-and-tested-by: Ravi Bangoria <ravi.bangoria@amd.com>
Link: https://lore.kernel.org/r/20250227222411.3490595-4-seanjc@google.com
Signed-off-by: Sean Christopherson <seanjc@google.com>
Signed-off-by: Mukesh-Ogare <OgreKumar.Mukesh@amd.com>
Signed-off-by: mohanasv2 <mohanasv@amd.com>