arm64: add cnp verification on Hisilicon erratum 162100125
virt inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBN3WI

------------------------------------------------------------------------

When errata 162100125 matches, only the MIDR value is checked to see
if the CPU matches. It is also important to check whether the CnP in
the MMFR2 register is enabled. If CnP is not enabled, this errata is
not required.

Signed-off-by: Jinqian Yang <yangjinqian1@huawei.com>
Signed-off-by: Zhou Wang <wangzhou1@hisilicon.com>