fs/resctrl: Introduce the resctrl_feat_type
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IC35SV

--------------------------------

In order to expand the new features of MPAM in the following patches, enum
resctrl_feat_type is newly added.

When the configuration is updated, the CDP type and configuration feature
will be traversed and updated, so the staged_config array used to deliver
the configuration is expanded into a two-dimensional array.

Signed-off-by: Zeng Heng <zengheng4@huawei.com>