x86/devicetree: Fix the ioapic interrupt type table
mainline inclusion
from mainline-v5.11-rc1
commit 2e730cb56b2cd1626fecaf23ef1537fb24721ef2
category: bugfix
bugzilla: https://gitee.com/src-openeuler/kernel/issues/IC9GHN

Reference: https://github.com/torvalds/linux/commit/2e730cb56b2cd1626fecaf23ef1537fb24721ef2

--------------------------------

The ioapic interrupt type table is wrong as it assumes that polarity in
IO/APIC context means active high when set. But the IO/APIC polarity is
working the other way round. This works because the ordering of the entries
is consistent with the device tree and the type information is not used by
the IO/APIC interrupt chip.

The whole trigger and polarity business of IO/APIC is misleading and the
corresponding constants which are defined as 0/1 are not used consistently
and are going to be removed.

Rename the type table members to 'is_level' and 'active_low' and adjust the
type information for consistency sake.

No functional change.

Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Signed-off-by: David Woodhouse <dwmw@amazon.co.uk>
Signed-off-by: Thomas Gleixner <tglx@linutronix.de>
Link: https://lore.kernel.org/r/20201024213535.443185-5-dwmw2@infradead.org
Signed-off-by: Liu Chao <liuchao173@huawe.com>