ucounts: reinplement rlimit with percpu_counter
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IC97W5

----------------------------------------

The will-it-scale test case signal1 [1] has been observed. and the test
results reveal that the signal sending system call lacks linearity.
To further investigate this issue, we initiated a series of tests by
launching varying numbers of dockers and closely monitored the throughput
of each individual docker. The detailed test outcomes are presented as
follows:

  | Dockers 	|1	|4	|8	|16	|32	|64	|
  | Throughput 	|380068	|353204	|308948	|306453	|180659	|129152	|

The data clearly demonstrates a discernible trend: as the quantity of
dockers increases, the throughput per container progressively declines.
In-depth analysis has identified the root cause of this performance
degradation. The ucounts module conducts statistics on rlimit, which
involves a significant number of atomic operations. These atomic
operations, when acting on the same variable, trigger a substantial number
of cache misses or remote accesses, ultimately resulting in a drop in
performance.

To address the above issues, this patch converts the atomic rlimit to a
percpu_counter. Summing up the percpu counters is expensive. To overcome
this, this patch modifies the conditions for freeing ucounts. Instead of
complex checks regarding whether a pending signal is the first or the last
one, the ucounts can now be freed only when both the refcount and the
rlimits are zero.

After the optimization, the performance data is shown below, demonstrating
that the throughput no longer declines as the number of Docker containers
increases:

  | Dockers 	|1	|4	|8	|16	|32	|64	|
  | Throughput 	|374737	|376377	|374814	|379284	|374950	|377509	|

[1] https://github.com/antonblanchard/will-it-scale/blob/master/tests/
Signed-off-by: Chen Ridong <chenridong@huawei.com>