crypto: hisilicon/qm - alloc power of 2 interrupts
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/IC9UQB
CVE: NA

----------------------------------------------------------------------

If the number of interrupts is not the power of 2,
some interrupts cannot be unmapd when the interrupts are released.
As a result, allocating interrupts again will fail. Therefore,
when allocating interrupts, change the number of interrupts to the
power of 2. For vf, irq vector 3 is reserved, request a
reserved interrupt, for pf, request abnormal irq, the redundant
check is deleted.

Fixes: dbdc1ec31fc0 ("crypto: hisilicon - add device error report through abnormal irq")
Signed-off-by: Weili Qian <qianweili@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>