crypto: hisilicon - enable error reporting again
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IC9UQB
CVE: NA

----------------------------------------------------------------------

When an error occurs on the device, an interrupt is reported.
When the firmware forwards the interrupt to the driver and masks the
error. If the driver does not enable error reporting when an error does
not need to be reset, the device does not report the error to the driver
when the error occurs again. Therefore, after the driver obtains the
information, the error reporting needs to be enabled again.

Signed-off-by: Weili Qian <qianweili@huawei.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>