sched: More flexible use of CPU quota when CPU is idle
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ICE7WC

--------------------------------

This feature allows users to use CPU quota more flexibly when CPU is
idle and it will cause the CPU quota to be exceeded. So, it cannot be
used in scenarios where there are strict restrictions on the use of the
CPU quota, such as some commercial scenarios that charge based on the
use of CPU quota.

Signed-off-by: Zheng Zucheng <zhengzucheng@huawei.com>
Signed-off-by: Liao Chang <liaochang1@huawei.com>
Signed-off-by: Cheng Yu <serein.chengyu@huawei.com>