sched/fair: reset idlest_cpu if not meet the conditions for preferred CPU
jingdong inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/I8PGAS

-----------------------------------------------------

We have set the preferred CPU, but when the load is too high,
the conditions for the preferred CPU are not met.
At this time, the idle CPU needs to be reset.

This will be corrected in the subsequent fallback process.
dynamic affinity should not break this logic.

Fixes: 2e1dfc02d115 ("sched: Adjust wakeup cpu range according CPU util dynamicly")
Signed-off-by: zhangwei123171 <zhangwei123171@jd.com>
Reviewed-by: zhaoxiaoqiang11 <zhaoxiaoqiang11@jd.com>
Signed-off-by: Cheng Yu <serein.chengyu@huawei.com>