arm64: Add hardware xcall framework support
In the svc exception handling process of the ARM64, some auxiliary
functions such as debug/trace and core functions such as sched/signal
need to be considered. As a result, the svc exception handling process
is very "lengthy". In some short syscalls (such as lseek(), getpid(),
etc.) handles the syscall function itself with a small percentage of
instructions, which also leads to inefficient execution of such
syscalls.

The idea of xcall can be regarded as a fast svc exception handling path.
In the processing process, only the context save and restore and
syscall functions are considered. It can be understood that xcall is a
fast syscall processing mechanism between vdso and normal syscalls.

The svc instruction with the immediate 0xffff is used to transfer xcall
information. At the synchronization exception handling entry of the
arm64 vectors, the ESR_EL1 register is used to distinguish xcall from
normal syscalls.

Signed-off-by: Jinjie Ruan <ruanjinjie@huawei.com>