bpf-rvi: cgroup: Add cgroup_rstat_flush_atomic() kfunc
hulk inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ICA1GK

--------------------------------

Restore cgroup_rstat_flush_atomic() that was removed by commit
0a2dc6ac3329 ("cgroup: remove cgroup_rstat_flush_atomic()") and expose
it to bpf.

This function is basically the same with cgroup_rstat_flush(), and the
later one is sleepable only in PREEMPT_RT case, which is not a supported
scenario of this feature.

Signed-off-by: GONG Ruiqi <gongruiqi1@huawei.com>