perf pmu-events: Fix filter 'ext' filed config incorrect
driver inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ICRSOQ

----------------------------------------------------------------------

The ext function for collecting L3D events is incorrectly enabled. As a result,
the collection result of L3C events is 0. The ext field is added to the filter.
The ext field is 1 for common L3 events, and the ext field is 2 for L3 ext events.

Signed-off-by: Qinxin Xia <xiaqinxin@huawei.com>
Signed-off-by: Qizhi Zhang <zhangqizhi3@h-partners.com>
Signed-off-by: JiangShui Yang <yangjiangshui@h-partners.com>