arm64: Fix actlr_el1 can not set problem in guest
hulk inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/release-management/issues/IBV2E4

--------------------------------

During the initialization of the vCPU, the vcpu->arch.hcr_el2 is set to
HCR_GUEST_FLAGS, which include HCR_TACR, so regardless of whether
the host has configured HCR_EL2, accessing ACTLR_EL1 in the guest
virtual machine will trap to EL2.

However, the current implementation of KVM ignores writes to ACTLR_EL1,
so it is currently not possible to set this register in the guest,
hardware xcall/xint can not take effect anyway, so modify the state of
actlr based on the written value in access_actlr().

Fixes: 7f2e02718bba ("arm64: entry: Support hardware xcall and xint")
Signed-off-by: Jinjie Ruan <ruanjinjie@huawei.com>