KVM: arm64: Expose Spec_FPACC to guests
virt inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IBN3WI

------------------------------------------------------------------------

920G: ID_AA64MMFR3_EL1.Spec_FPACC=0x1
920B/C: ID_AA64MMFR3_EL1.Spec_FPACC=0x0
The live migration between 920B/C/G need to change the value of Spec_FPACC,
so it is exposed to the guest. Closing the FPACC on the guest does not
directly affect the FPACC status on the host.

Sign-off-by: Jinqian Yang <yangjinqian1@huawei.com>