drm/hisilicon/hibmc: fix the VGA output has no modes and EDID problem
driver inclusion
category: bugfix
bugzilla: https://gitee.com/openeuler/kernel/issues/ID3LNC
CVE: NA

--------------------------------------------------------------

When the parameter video-VGA@640x480 was write in the GRUB, the
DRM_FORCE_<foo> in VGA connector will be ON. In that case, detect_ctx()
won't be call to detect the VGA, so the VGA phys_state won't be updated.
There is a problem that the state will always be disconnect even if the
VGA device plugged in.

Fixes: 2a581acede34 ("drm/hisilicon/hibmc: hibmc-drm bugfix for DP")
Signed-off-by: Baihan Li <libaihan@huawei.com>