usb: pci-quirks: Fix TBT USB Enumeration Long Blocking Issue
zhaoxin inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/IDAQZI
CVE: NA

-------------------

At two Type-C ports, a set of devices are cascaded respectively,
including a TBT4 dock, a TBT3 storage device, and a USB2.0 device.
During the enumeration process, due to the addition of serial port
identification, the U3 hub in the U3 dock slows down, and it is prone to
triggering the patch of the FW U3 hub to switch the Vbus. At this time,
the PCIe hierarchy in the secondary TBT3 storage device may still be in
the process of enumeration, which may lead to unpredictable disconnections
in PCIe/NVMe/XHCI.
The corresponding driver does not expect disconnections to occur during
the enumeration when enumerating these controllers. During xHCI handoff
and other CNR periods, the driver may loop - read MMIO up to 500000 times.
Since the controller is disconnected, each MMIO operation will experience
a 30-millisecond timeout. This may result in a maximum waiting time of
15000 seconds in this loop, which may trigger more deadlocks and
ultimately cause the system to seem unresponsive.
Previously, some cases have been found where the system experiences a
black screen or freezes because some CPU cores remain in the NVMe and PCIe
drivers for an extended period.

Signed-off-by: leoliu-oc <leoliu-oc@zhaoxin.com>