Compiler: Add clang's PGO support for kernel.
compile inclusion
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ID5H9D

--------------------------------

Enable the use of clang's Profile-Guided Optimization[1]. To generate a
profile, the kernel is instrumented with PGO counters, a representative
workload is run, and the raw profile data is collected from
/sys/kernel/debug/pgo/profraw.

The raw profile data must be processed by clang's "llvm-profdata" tool
before it can be used during recompilation:

  $ cp /sys/kernel/debug/pgo/profraw vmlinux.profraw
  $ llvm-profdata merge --output=vmlinux.profdata vmlinux.profraw

Multiple raw profiles may be merged during this step.

The data can now be used by the compiler:

  $ make LLVM=1 KCFLAGS=-fprofile-use=vmlinux.profdata ...

This initial submission is restricted to x86, as that's the platform we
know works. This restriction can be lifted once other platforms have
been verified to work with PGO.

Note that this method of profiling the kernel is clang-native, unlike
the clang support in kernel/gcov.

[1] https://clang.llvm.org/docs/UsersManual.html#profile-guided-opti
mization

[Reference]
https://groups.google.com/g/clang-built-linux/c/85eKdM9_Jeg/m/WDqPmi
8DDwAJ
https://gitee.com/openeuler/kernel/pulls/9294

Signed-off-by: gaochong <gaochong@kylinsec.com.cn>