xfs: add xfs_calc_atomic_write_unit_max()
mainline inclusion
from mainline-v6.15-rc2
commit 0c438dcc31504bf4f50b20dc52f8f5ca7fab53e2
category: feature
bugzilla: https://gitee.com/openeuler/kernel/issues/ID2HML
CVE: NA

Reference: https://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git/commit/?id=0c438dcc31504bf4f50b20dc52f8f5ca7fab53e2

--------------------------------

Now that CoW-based atomic writes are supported, update the max size of an
atomic write for the data device.

The limit of a CoW-based atomic write will be the limit of the number of
logitems which can fit into a single transaction.

In addition, the max atomic write size needs to be aligned to the agsize.
Limit the size of atomic writes to the greatest power-of-two factor of the
agsize so that allocations for an atomic write will always be aligned
compatibly with the alignment requirements of the storage.

Function xfs_atomic_write_logitems() is added to find the limit the number
of log items which can fit in a single transaction.

Amend the max atomic write computation to create a new transaction
reservation type, and compute the maximum size of an atomic write
completion (in fsblocks) based on this new transaction reservation.
Initially, tr_atomic_write is a clone of tr_itruncate, which provides a
reasonable level of parallelism.  In the next patch, we'll add a mount
option so that sysadmins can configure their own limits.

[djwong: use a new reservation type for atomic write ioends, refactor
group limit calculations]

Reviewed-by: Darrick J. Wong <djwong@kernel.org>
Signed-off-by: Darrick J. Wong <djwong@kernel.org>
[jpg: rounddown power-of-2 always]
Reviewed-by: Christoph Hellwig <hch@lst.de>
Signed-off-by: John Garry <john.g.garry@oracle.com>
Conflicts:
	fs/xfs/libxfs/xfs_trans_resv.c
	fs/xfs/xfs_mount.c
	fs/xfs/xfs_mount.h
	fs/xfs/xfs_reflink.h
	fs/xfs/xfs_trace.h
Signed-off-by: Long Li <leo.lilong@huawei.com>