mfs: Add basic metadata operation for MFS
Offering: HULK
hulk inclusion
category: feature
bugzilla: NA

--------------------------------

It is easy to add metadata operation in MFS. In MFS, metadata is
from the lower layer, so we just need to pass the requests to the
lower inode/dentry and adjust the reference.

In remote mode, the cache layer may be empty, so we should create
the target file when lookup.

Signed-off-by: Huang Xiaojia <huangxiaojia2@huawei.com>
Signed-off-by: Hongbo Li <lihongbo22@huawei.com>