mfs: Add basic data operation for MFS
Offering: HULK
hulk inclusion
category: feature
bugzilla: NA

--------------------------------

It is easy to add data operation in MFS include file and directory.
In MFS, data is from the cache layer, so we just need to pass the
requests to the cache layer.

In this step, we just transfer the read io to the under file system.

Signed-off-by: Huang Xiaojia <huangxiaojia2@huawei.com>
Signed-off-by: Hongbo Li <lihongbo22@huawei.com>