mfs: Add basic cache framework: check method and control structure
Offering: HULK
hulk inclusion
category: feature
bugzilla: NA

--------------------------------

When the read request miss in cache (memory or disk), MFS will
post miss event to userspace. And we should use different way
to check. For memory, we just need to check the target range
whether in page cache or not. And for disk (remote mode), we
should use SEEK_DATA and SEEK_HOLE to check the target range.

In this patch, we introduce the basic structure for events that
is @mfs_syncer. In remote mode, the events are synchronous, so
all the io segments from the same reading range will wait on the
same syncer object. Also the @mfs_cache_object is introduced to
imply the cache object in cache layer. @mfs_cache_object is the
controller for MFS's data. It builds the connection between MFS's
data and the events.

Signed-off-by: Huang Xiaojia <huangxiaojia2@huawei.com>
Signed-off-by: Hongbo Li <lihongbo22@huawei.com>