mfs: Add basic events framework
Offering: HULK
hulk inclusion
category: feature
bugzilla: NA

--------------------------------

All events posted during data miss will link into the xarray
structure. The @mfs_event is introduced to represent the
target miss event. For now, only read event should be considered.

The message header in the event is encoded in @mfs_msg, the user
use the fd as file handler to control the cache in MFS. The core
structure @mfs_cache_object is kept in the inode's priviate data
at the time that the inode was allocated.

Signed-off-by: Huang Xiaojia <huangxiaojia2@huawei.com>
Signed-off-by: Hongbo Li <lihongbo22@huawei.com>