mfs: Add user command for handling events
Offering: HULK
hulk inclusion
category: feature
bugzilla: NA

--------------------------------

After the user obtains the events, the following action should
be done. For async events, MFS provides the readahead command
for any range of the file. And this is different with the default
readahead system call such as posix_fadvise which limits the io
size. For sync events, the user should fetch the raw data from
remote storage server and write to fd. After handling the target
event, the user should execute the IOC_DONE command to notify
the sync event has been handled which will wakeup the waiting
io in kernel.

Signed-off-by: Huang Xiaojia <huangxiaojia2@huawei.com>
Signed-off-by: Hongbo Li <lihongbo22@huawei.com>