mfs: Fix the UAF problems on syncer object
Offering: HULK
hulk inclusion
category: bugfix
bugzilla: NA

------------------

In remote mode, undo events are placed on the syncer list. The syncer
object is a stack variable and will be released if the process is
interrupted or the syncer is woken up during the closing of the device's
file descriptor. When the daemon exits, it calls `dev_release` and wakes
up all events. Once the syncer is marked as completed, accessing it will
lead to UAF vulnerability. Therefore, the syncer must never be used after
its events have been completed.

Additionally, `mfs_cancel_syncer_events` is called when a user process
terminates. There are scenarios where event completion may coincide with
other cleanup tasks, which can result in invalid event references. To
prevent this, the `xa_lock` is used to protect against the race condition.

Fixes: 3236f3b541d1 ("[Huawei] mfs: Add communication devie and event acquisition operations")
Signed-off-by: Hongbo Li <lihongbo22@huawei.com>